requirements.txt
bs4==0.0.1
moviepy==1.0.0
python-dotenv==0.10.1
pytube==9.4.0
requests==2.21.0
tqdm==4.31.1
git+https://github.com/tucraig/tweepy.git
